pyproject.toml
[tool.pdm]
[tool.pdm.dev-dependencies]
dev = [
    "isort>=5.12.0",
]

[project]
name = ""
version = ""
description = ""
authors = [
    {name = "Miguel", email = "Miguelbochibarro@gmail.com"},
]
dependencies = [
    "django>=4.1.7",
    "djangorestframework>=3.14.0",
    "setuptools>=68.1.2",
    "python-magic>=0.4.27",
    "Pillow>=10.0.0",
]
requires-python = ">=3.10"
license = {text = "MIT"}
